smartmeal/cafeteria-predictor/requirements.txt
# Uncomment to enable Azure Monitor OpenTelemetry
# Ref: aka.ms/functions-azure-monitor-python 
# azure-monitor-opentelemetry 

azure-functions
pandas
scikit-learn
joblib
azure-storage-blob
